add btsg-ft-factory to cw-orch setup
