+++ packages/cw-orch/src/external/mod.rs
@@ -5,6 +5,7 @@ mod migrator;
 mod payroll_factory;
 mod token_swap;
 mod tokenfactory_issuer;
+mod btsg_ft_factory;
 
 pub use admin_factory::DaoExternalAdminFactory;
 pub use cw721_roles::DaoExternalCw721Roles;
@@ -13,3 +14,4 @@ pub use migrator::DaoExternalMigrator;
 pub use payroll_factory::DaoExternalPayrollFactory;
 pub use token_swap::DaoExternalTokenSwap;
 pub use tokenfactory_issuer::DaoExternalTokenfactoryIssuer;
+pub use btsg_ft_factory::DaoExternalFantokenFactory;
