+++ scripts/src/lib.rs
@@ -2,9 +2,9 @@
 fn main() {}
 
 mod dao;
+mod external;
 pub use dao::*;
+pub use external::*;
 
-#[cfg(test)]
-mod deploy;
 #[cfg(test)]
 mod tests;
