+++ scripts/src/tests/external/mod.rs
@@ -5,3 +5,4 @@ pub mod token_swap;
 pub mod tokenfactory_issuer;
 pub mod vesting;
 pub mod dao_migration;
+pub mod btsg_ft_factory;
