+++ scripts/src/tests/mod.rs
@@ -1,7 +1,7 @@
 mod external;
+mod deploy;
 
 pub(crate) const PREFIX: &str = "mock";
 pub(crate) const ADMIN: &str = "admin";
 pub(crate) const DENOM: &str = "juno";
-
 pub(crate) const DAO1: &str = "dao1";
